+++ clove/CMakeLists.txt
@@ -204,7 +204,7 @@ target_link_libraries(
 		#Systems
 		CloveAI
 		CloveECS
-		CloveReflection
+		CloveSerialisation
 		CloveStateMachine
 		 
 		#Utilities
