+++ clove/components/systems/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(ai)
 add_subdirectory(ecs)
-add_subdirectory(reflection)
+add_subdirectory(serialisation)
 add_subdirectory(state_machine)
