+++ clove/components/systems/serialisation/source/Serialiser.cpp
@@ -1,0 +1,5 @@
+#include "Clove/Serialisation/Serialiser.hpp"
+
+namespace garlic::clove {
+
+}
